# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/nurses": {
            "get": {
                "description": "list nurse entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List nurse entities",
                "operationId": "list-nurse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Nurse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create nurse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create nurse",
                "operationId": "create-nurse",
                "parameters": [
                    {
                        "description": "Nurse entity",
                        "name": "nurse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Nurse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Nurse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/nurses/{id}": {
            "get": {
                "description": "get nurse by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a nurse entity by ID",
                "operationId": "get-nurse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Nurse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Nurse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update nurse by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a nurse entity by ID",
                "operationId": "update-nurse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "nurse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "nurse entity",
                        "name": "nurse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Nurse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Nurse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get nurse by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a nurse entity by ID",
                "operationId": "delete-nurse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "nurse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Examinationroom": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ExaminationroomQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ExaminationroomEdges"
                },
                "examinationroom_name": {
                    "description": "ExaminationroomName holds the value of the \"examinationroom_name\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ExaminationroomEdges": {
            "type": "object",
            "properties": {
                "examinationroomOperativerecord": {
                    "description": "ExaminationroomOperativerecord holds the value of the Examinationroom_Operativerecord edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Operativerecord"
                    }
                }
            }
        },
        "ent.Nurse": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the NurseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.NurseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "nurse_Email": {
                    "description": "NurseEmail holds the value of the \"nurse_Email\" field.",
                    "type": "string"
                },
                "nurse_Name": {
                    "description": "NurseName holds the value of the \"nurse_Name\" field.",
                    "type": "string"
                },
                "nurse_Password": {
                    "description": "NursePassword holds the value of the \"nurse_Password\" field.",
                    "type": "string"
                },
                "nurse_Tel": {
                    "description": "NurseTel holds the value of the \"nurse_Tel\" field.",
                    "type": "string"
                }
            }
        },
        "ent.NurseEdges": {
            "type": "object",
            "properties": {
                "nurseOperativerecord": {
                    "description": "NurseOperativerecord holds the value of the Nurse_Operativerecord edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Operativerecord"
                    }
                }
            }
        },
        "ent.Operative": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the OperativeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.OperativeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "operative_Name": {
                    "description": "OperativeName holds the value of the \"operative_Name\" field.",
                    "type": "string"
                },
                "operative_Type": {
                    "description": "OperativeType holds the value of the \"operative_Type\" field.",
                    "type": "string"
                }
            }
        },
        "ent.OperativeEdges": {
            "type": "object",
            "properties": {
                "operativeOperativerecord": {
                    "description": "OperativeOperativerecord holds the value of the Operative_Operativerecord edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Operativerecord"
                    }
                }
            }
        },
        "ent.Operativerecord": {
            "type": "object",
            "properties": {
                "OperativeTime": {
                    "description": "OperativeTime holds the value of the \"OperativeTime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the OperativerecordQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.OperativerecordEdges"
                },
                "examinationroom_id": {
                    "type": "integer"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "nurse_id": {
                    "type": "integer"
                },
                "operative_id": {
                    "type": "integer"
                },
                "tool_id": {
                    "type": "integer"
                }
            }
        },
        "ent.OperativerecordEdges": {
            "type": "object",
            "properties": {
                "examinationroom": {
                    "description": "Examinationroom holds the value of the Examinationroom edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Examinationroom"
                },
                "nurse": {
                    "description": "Nurse holds the value of the Nurse edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Nurse"
                },
                "operative": {
                    "description": "Operative holds the value of the Operative edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Operative"
                },
                "tool": {
                    "description": "Tool holds the value of the Tool edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Tool"
                }
            }
        },
        "ent.Tool": {
            "type": "object",
            "properties": {
                "Tool_Amount": {
                    "description": "ToolAmount holds the value of the \"Tool_Amount\" field.",
                    "type": "string"
                },
                "Tool_Name": {
                    "description": "ToolName holds the value of the \"Tool_Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ToolQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ToolEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ToolEdges": {
            "type": "object",
            "properties": {
                "toolOperativerecord": {
                    "description": "ToolOperativerecord holds the value of the Tool_Operativerecord edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Operativerecord"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}